Fix timing after 1 million seconds and only let agents eat when moving forwards

+++ src/world/time_controller.rs
@@ -17,7 +17,7 @@ impl TimeController {
         }
     }
 
-    pub fn goto(&mut self, params: &Params, world: &mut World, new_time: f32) {
+    pub fn goto(&mut self, params: &Params, world: &mut World, new_time: f64) {
         if new_time < world.time {
             *world = World::new(params);
         }
